Added Backup::Exception::CommandNotFound when command line utilities cannot be found and the OS returns a 32512 error.
Added a method for Backup::CLI ( #raise_if_command_not_found! ) that handles throwing exceptions when Backup::CLI runs a command of which the main utility could not be found (e.g. mysqldump, mongodump).

The Logger now has a Logger.normal method that simply does a regular "puts" command, but at the same time, logs to the backup.log for debugging purposes. This is now used when exceptions are thrown since it would previously not log exceptions to the backup.log which made it harder to debug issues.

Backup now handles any time of exception that may be raised and gives a more detailed output as well. Rather than using ENV variables, Backup now uses Ruby's built-in constants since the ENV variables sometimes appeared to not display particular details. Now the RUBY_PLATFORM is also included for better debugging. Also the exception's class name will be displayed from now on.

Finally, The email notifier template has been updated with all these details, so now detailed information will be displayed in the console, in the backup.log and in the email notifications to make debugging and tracking the issues easier.