Resolve monkeypatching by explicitly using Test* classes
This prevents the entire spec method from failing on Gzip, or other truly existing classes
only when all the specs are run. i.e. rspec spec/

Using a mock for the new method for Backup::Archive, safely reimplements it to just return
true, which satisfies the specification for this spec.