use `-P` option to preserve leading '/' in Archive
This is to suppress the warning:
  $ tar: Removing leading `/' from member names

When extracting the resulting archive, tar will still extract these
paths as relative to the working directory, even with the leading '/'
preserved for the paths in the archive.
Unless you specify the `-P` option when extracting the archive,
tar will remove the leading '/' from member names and issue this
warning at that time.