update VM spec helpers
Update Package so it can find the package files for a specific Model.
This allows a model to be run multiple times, which is required to test
cycling. This also ensures the package was in fact stored in the
intended remote_path.

Don't create a Tarfile unless Package inspection methods are used. This
prevents a split package from being rejoined within the remote_path,
which would disrupt some cycling routines.