Add Server-Side encryption option to S3
Adds an attribute that can be used with S3 storage to enable Server-Side
encryption for files uploaded to S3. This is done by include the header
`x-amz-server-side-encryption` as outlined in the Amazon S3 SDK:

http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingServerSideEncryption.html

Also included in this commit are tests for the enhancements, as well as
an updated Readme with an example of turning on Server-Side encryption
for S3.

Related Issue: #408