Require time library. Closes #727
In normal envs the Time library isn't loaded unless explicitly called.
It works in the test environment, because the timecop gem is loaded and
this requires the time library for us.

Now explictly requiring time and adding a spec for the method that
depends on it (since I forgot to write it when adding it).