Don't store project files on docker image
This only prepares the build environment on the docker image, but does
not create a production image which could be deployed for an app.

It uses volumes and a custom `BUNDLE_PATH` to store the gems in the
bundle which can be used between updates to the bundle itself. If a gem
version changes it would not have to reinstall all the gems and build
the image from that point, it would be able to reuse the bundler cache
(in the mounted volume) and only require an update to the changed gems.
Overal, should speed up the build process for the docker test
environment.

Added a `bin/docker_test` file that performs tasks within the docker
container.

Also made building the docker image mandatory when running the `docker`
namespaced rake files, so you can't end up with an out-of-sync state
where the docker image has changed, but hasn't been rebuild so a
developer would be testing against a stale image.